LPS-195397 Inline DLUploadFileEntryHandler in UploadFileEntryMVCActionCommand
